requirements.txt
discord.py==2.3.2
mysql-connector-python
python-dotenv==1.0.0
audioop-lts